Perform package squash
Ansible 2.7 handles looping with package managers differently. [1]
It used to perform list squashing when the loop detected it was being
combined with a package manager module that supported installing the
packages in a single transaction. Because this is no longer the case
it's better to give a list of packages to the package manager module.

[1]: https://docs.ansible.com/ansible/devel/porting_guides/porting_guide_2.7.html#using-a-loop-on-a-package-module-via-squash-actions